beta 7

+++ q_a/main.py
@@ -1,5 +1,5 @@
 from q_a import app
-from flask import render_template, Markup, request, url_for, redirect
+from flask import render_template, Markup, request, url_for, redirect, session
 from q_a.models import Post, Tags, db
 from q_a.supplement import Amend, Check
 
@@ -17,9 +17,10 @@ def main(page=1):
                                Tags=Tags,
                                posts=news)
     elif request.method == 'POST':
-        if request.form.get('delete_post'):
+        if request.form.get('delete_post') and session.get('status') == 2:
             db.session.delete(Post.query.get(request.form.get('delete_post')))
             db.session.commit()
             return Amend.flash('Пост удалён.', 'success', url_for('main'))
-        elif request.form.get('edit_post'):
+        elif request.form.get('edit_post') and session.get('status') == 2:
             return redirect(url_for('edit_post', post_id=request.form.get('edit_post')))
+        return Check.status()
